helpdesk.tickets-service/src/main/java/innowise/microservice/helpdesk/ticketsservice/controller/FeedbackController.java
package innowise.microservice.helpdesk.ticketsservice.controller;

import innowise.microservice.helpdesk.ticketsservice.dto.FeedbackDTO;
import innowise.microservice.helpdesk.ticketsservice.entity.Feedback;
import innowise.microservice.helpdesk.ticketsservice.entity.User;
import innowise.microservice.helpdesk.ticketsservice.services.FeedbackService;
import lombok.AllArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.Optional;

@RestController
@AllArgsConstructor
@RequestMapping("/helpdesk-service/feedback")
public class FeedbackController {

    private final FeedbackService feedbackService;

    @PostMapping
    public ResponseEntity<String> createFeedback(@RequestBody FeedbackDTO feedbackDTO, @AuthenticationPrincipal User creator) {
        feedbackService.createFeedback(feedbackDTO, creator);
        return ResponseEntity.ok("Feedback created successfully");
    }

    @GetMapping("/{id}")
    public ResponseEntity<Optional<Feedback>> getFeedback(@PathVariable("id") long id) {
        Optional<Feedback> feedback = feedbackService.getFeedbackById(id);
        return new ResponseEntity<>(feedback, HttpStatus.OK);
    }
}